Fixing Fisher Yates and the sampling method
._FisherYates wasn't allowing steps=n-1, or rather, it was, it's just it was returning the whole array when that happened because n-1 steps is a complete shuffle. If steps is n-1 now it returns a list of the last n-1 elements, which I think is the expected behaviour if you explicitly call steps=n-1. 

Also added an 'orderMatters' parameter to .sample, by default true, because the entropy-optimization trick returns a list where the elements are in their original order and that might be unexpected behaviour depending on how you use this. So if you just want a random sample as if selecting one handful of elements at once rather than sequentially, do set orderMatters to false. In the future I might actually simply divide this into two separate methods, one for "handful" sampling, another for when you are sequentially choosing elements.