Default seed added, updated sample and the readme
If reseed is called without a seed, it'll use a string of 0s to instantiate the ChaChaRand object. If the constructor is called without a seed, it'll do the same but also warn that we aren't using some random value (since some may expect the default to be taken from the system's time or any other such standard default seeding method).

README also got an update with some minimal examples.