MAX_SAFE_BITS was ill defined
Although coincidentally correct, because log2(2^53-1) is so close to 53 that JS just approximated to 53 (it still does this, even if you put it inside a Math.floor, which *should* give 52 instead). So I'm just sticking with a hardcoded 53 for now, for my own peace of mind.

Also, fixing some comments and an error message.