Fix of choices, more seed format validation and other lil' changes.
Choices had a typo where it was doing randUint(n) instead of n-1. 
It was also unnecessarily throwing error on choicesSize > n-1.

reseed now validates that the given seed is in hex format.

myRange and fromHexString moved to be static ChaChaGen functions... also
ChaChaRand was renamed to ChaChaGen (so it doesn't coincide with Kudelski Security's RNG's name).  (...I'm thiiis close to rerenaming it "ChaChaChance").

Sample now throws an error if no sample size is provided.